+++ docs/tailwind.config.ts
@@ -1,6 +1,6 @@
 import { Config } from 'tailwindcss/types/config';
 
-import { preset } from './theme/preset';
+import { preset } from '@marigold/theme-docs/preset';
 
 const config: Config = {
   content: [
