+++ themes/theme-docs/README.md
@@ -1,0 +1,3 @@
+# `@marigold/theme-docs`
+
+> Marigol's theme used for the docs website
