+++ themes/theme-docs/package.json
@@ -1,0 +1,60 @@
+{
+  "name": "@marigold/theme-docs",
+  "version": "26.1.9",
+  "description": "Marigold Docs Theme",
+  "license": "MIT",
+  "keywords": [
+    "marigold",
+    "docs",
+    "theme",
+    "styles"
+  ],
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "require": "./dist/index.js",
+      "import": "./dist/index.mjs",
+      "types": "./dist/index.d.ts"
+    },
+    "./preset": {
+      "require": "./dist/preset.js",
+      "import": "./dist/preset.mjs",
+      "types": "./dist/preset.d.ts"
+    },
+    "./styles.css": "./dist/styles.css"
+  },
+  "typesVersions": {
+    "*": {
+      "*": [
+        "./dist/*"
+      ]
+    }
+  },
+  "sideEffects": false,
+  "files": [
+    "dist"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/marigold-ui/marigold",
+    "directory": "themes/theme-docs"
+  },
+  "devDependencies": {
+    "@marigold/tsconfig": "workspace:*",
+    "autoprefixer": "10.4.19",
+    "cssnano": "6.1.2",
+    "postcss": "8.4.35",
+    "tailwindcss": "3.4.1",
+    "tsup": "8.0.2"
+  },
+  "dependencies": {
+    "@marigold/components": "workspace:*",
+    "@marigold/system": "workspace:*",
+    "@marigold/theme-preset": "workspace:*"
+  },
+  "scripts": {
+    "build": "NODE_ENV=production tsup && tailwindcss -o dist/styles.css"
+  }
+}
