+++ themes/theme-docs/src/components/Dialog.styles.ts
