+++ themes/theme-docs/src/components/Table.styles.ts
