+++ themes/theme-docs/src/components/Tabs.styles.ts
