+++ themes/theme-docs/src/components/index.ts
