+++ themes/theme-docs/src/root.ts
