+++ themes/theme-docs/tsconfig.build.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@marigold/tsconfig/base"
+}
