+++ themes/theme-docs/tsup.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from 'tsup';
+
+export default defineConfig({
+  entry: ['src/index.ts', 'src/preset.ts'],
+  format: ['esm', 'cjs'],
+  tsconfig: './tsconfig.build.json',
+  dts: true,
+  sourcemap: true,
+  clean: process.env.NODE_ENV === 'production',
+});
